+++ demo.js
@@ -5,7 +5,7 @@ import 'lexgui/components/audio.js';
 
 window.LX = LX;
 
-let area = LX.init( { strictViewport: false } );
+const area = LX.init( { strictViewport: false, rootClass: "wrapper" } );
 
 // Menubar
 {
@@ -40,12 +40,12 @@ let area = LX.init( { strictViewport: false } );
 
 // Header
 {
-    const header = LX.makeContainer( [ null, "auto" ], "col border gap-2 p-8" );
+    const header = LX.makeContainer( [ null, "auto" ], "flex flex-col border border-l-0 border-r-0 gap-2 p-8" );
     
     header.innerHTML = `
         <a>Get started with LexGUI.js</a>
         <h1>Build your application interface</h1>
-        <p style="max-width:32rem">A set of beautifully-designed, accessible widgets and components. 
+        <p class="font-light" style="max-width:32rem">A set of beautifully-designed, accessible widgets and components.
         No complex frameworks. Pure JavaScript and CSS. Open Source.</p>
     `;
     
@@ -59,8 +59,141 @@ let area = LX.init( { strictViewport: false } );
 
     // Mail
     {
-        const mailContainer = LX.makeContainer( [ null, "auto" ], "col bg-primary border rounded-lg p-6" );
+        const mailContainer = LX.makeContainer( [ null, "800px" ], "flex flex-col bg-primary border rounded-lg" );
         tabs.add( "Mail", mailContainer, { selected: true } );
+
+        const mailArea = new LX.Area({ className: "rounded-lg" });
+        mailContainer.appendChild( mailArea.root );
+        // const customHeader = document.createElement('div');
+        // customHeader.innerHTML = "Custom simple header";
+        const badgeClass = "ml-auto no-bg font-medium";
+
+        const sidebar = mailArea.addSidebar( m => {
+            m.add( "Inbox", { icon: "inbox", content: LX.badge("128", badgeClass, { asElement: true }) } );
+            m.add( "Drafts", { icon: "file", content: LX.badge("9", badgeClass, { asElement: true }) } );
+            m.add( "Sent", { icon: "paper-plane" } );
+            m.add( "Junk", { icon: "box-archive", content: LX.badge("23", badgeClass, { asElement: true }) } );
+            m.add( "Trash", { icon: "trash-can" } );
+            m.add( "Archive", { icon: "box-archive" } );
+            m.separator();
+            m.add( "Social", { icon: "user", content: LX.badge("972", badgeClass, { asElement: true }) } );
+            m.add( "Updates", { icon: "circle-info", content: LX.badge("342", badgeClass, { asElement: true }) } );
+            m.add( "Forums", { icon: "comments", content: LX.badge("96", badgeClass, { asElement: true }) } );
+            m.add( "Shopping ", { icon: "shopping-cart" } );
+            m.add( "Promotions", { icon: "box-archive", content: LX.badge("21", badgeClass, { asElement: true }) } );
+        }, { 
+            // collapseToIcons: false,
+            className: "border border-l-0 border-t-0 border-b-0",
+            parentClass: "rounded-lg",
+            headerTitle: "jxarco",
+            headerSubtitle: "alexroco.30@gmail.com",
+            headerImage: "https://raw.githubusercontent.com/jxarco/lexgui.js/refs/heads/master/images/icon.png",
+            // header: customHeader,
+            skipFooter: true,
+            onHeaderPressed: (e, element) => {
+                new LX.DropdownMenu( element, [
+                    "My Account",
+                    null,
+                    { name: "Profile", short: "P", icon: "user" },
+                    { name: "Billing", disabled: true, icon: "credit-card" },
+                    { name: "Settings", short: "S" },
+                    null,
+                    { name: "Team" },
+                    { name: "Invite users", icon: "search" },
+                    null,
+                    { name: "Github", icon: "github" },
+                    { name: "Support", submenu: [
+                        { name: "Email", icon: "envelope" },
+                        { name: "Message", submenu: [
+                            { name: "Whatsapp" },
+                            { name: "iMessage" },
+                        ]},
+                    ]  }
+                ], { side: "right", align: "end" });
+            }
+        });
+
+        const inboxArea = sidebar.siblingArea;
+        inboxArea.root.classList.add( "rounded-lg" );
+
+        var [ left, right ] = inboxArea.split({ sizes:["50%","50%"] });
+
+        // Manage Inbox
+        {
+            const inboxTabs = left.addTabs({ parentClass: "p-3 items-end border border-t-0 border-l-0 border-r-0", sizes: [ "auto", "auto" ], float: "end" });
+
+            const allMailContainer = LX.makeContainer( [ "100%", "100%" ], "flex flex-col" );
+            inboxTabs.add( "All mail", allMailContainer, { selected: true } );
+
+            // Filter
+            {
+                const allMailFilter = LX.makeContainer( [ "100%", "64px" ], "p-2" );
+                allMailContainer.appendChild( allMailFilter );
+                const filterInput = new LX.TextInput(null, "", null, 
+                    { inputClass: "outline", width: "100%", icon: "fa fa-magnifying-glass", placeholder: "Search..." } 
+                );
+                allMailFilter.appendChild( filterInput.root );
+            }
+
+            // Content
+            {
+                const allMailContent = LX.makeContainer( [ "100%", "calc(100% - 64px)" ], "flex flex-col p-4 gap-2 overflow-scroll" );
+                allMailContainer.appendChild( allMailContent );
+
+                const addMail = ( sender, subject, content, date, tags ) => {
+                    const msgContent = LX.makeContainer( [ "100%", "auto" ], 
+                        "flex flex-col border p-3 rounded-lg gap-2 select-none" );
+                    allMailContent.appendChild( msgContent );
+
+                    // Name, subject, date
+                    {
+                        const msgInfo = LX.makeContainer( [ "100%", "auto" ], "flex flex-col gap-0.5" );
+                        msgContent.appendChild( msgInfo );
+
+                        const msgNameDate = LX.makeContainer( [ "100%", "auto" ], "flex flex-row" );
+                        msgInfo.appendChild( msgNameDate );
+
+                        // Name + Date
+                        {
+                            const msgName = LX.makeContainer( [ "auto", "auto" ], "font-semibold text-md" );
+                            msgName.innerHTML = sender;
+                            msgNameDate.appendChild( msgName );
+
+                            const msgDate = LX.makeContainer( [ "auto", "auto" ], "fg-tertiary text-sm ml-auto self-center" );
+                            msgDate.innerHTML = date;
+                            msgNameDate.appendChild( msgDate );
+                        }
+
+                        const msgSubject = LX.makeContainer( [ "100%", "auto" ], "font-semibold text-sm" );
+                        msgSubject.innerHTML = subject;
+                        msgInfo.appendChild( msgSubject );
+                    }
+                    const msgText = LX.makeContainer( [ "100%", "auto" ], "text-sm line-clamp-2 fg-tertiary" );
+                    msgText.innerHTML = content;
+                    msgContent.appendChild( msgText );
+                    const msgTags = LX.makeContainer( [ "100%", "auto" ], "flex flex-row gap-0.5 font-semibold" );
+                    for( const tag of tags )
+                    {
+                        msgTags.appendChild( LX.badge( tag, "sm", { asElement: true } ) );
+                    }
+                    msgContent.appendChild( msgTags );
+                };
+
+                LX.requestJSON( "data/example_mail_data.json", (data) => {
+                    for( const item of data )
+                    {
+                        addMail( item.name, item.subject, item.content, item.date, item.tags );
+                    }
+                } )
+            }
+
+            inboxTabs.add( "Unread", document.createElement('div'));
+        }
+
+        // Manage Message Preview
+        {
+            right.root.className += " rounded-lg";
+        }
     }
 
     // Tasks