+++ changelog.md
@@ -5,7 +5,9 @@
 Widgets:
 - Added support for disabling resize in TextArea Widget `options.resize: false`.
 - Added support for using LX.ICONS on Button Widgets.
+- Tooltips supported in Button Widget `options.tooltip`. Uses either `options.title` or `name` as content.
 
+Started work on tooltips `LX.asTooltip(trigger, content)`.
 Added support for adding a class to LexGUI root on init `options.rootClass`.
 Added support for adding a class to Sidebar/Menubar areas on `Area.addSidebar` and `Area.addMenubar`.
 Added `options.asElement` to return `LX.badge` as HTMLElement instead of html string.